projects/evaluation/pyproject.toml
[project]
name = "owa-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "owa-core",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
]

[tool.uv.sources]
owa-core = { path = "../owa-core", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["owa"]
